Get cities from redux
